Ignore file-upload meteor-method by ostrio:files
The package ostrio:files registers meteor-method calls starting by "_FilesCollectionWrite_" followed by the name of the file-collection. These method-calls are used to upload files to the server. Monitoring these methods (which come in rapidly as we want to upload the file fastest possible) would result in sending the full file in a EJSON encoded way also to the monitoring system.